fix(azure-devops) utils to misc
