+++ integrations/azure-devops/azure_devops/misc.py
@@ -1,5 +1,16 @@
+from enum import StrEnum
 from datetime import datetime, timedelta
 from typing import Any
+
+
+class Kind(StrEnum):
+    REPOSITORY = "repository"
+    REPOSITORY_POLICY = "repository-policy"
+    PULL_REQUEST = "pull-request"
+    PIPELINE = "pipeline"
+    MEMBER = "member"
+    TEAM = "team"
+    PROJECT = "project"
 
 
 PULL_REQUEST_SEARCH_CRITERIA: list[dict[str, Any]] = [
