--- integrations/azure-devops/azure_devops/utils.py
@@ -1,11 +1,0 @@
-from enum import StrEnum
-
-
-class Kind(StrEnum):
-    REPOSITORY = "repository"
-    REPOSITORY_POLICY = "repository-policy"
-    PULL_REQUEST = "pull-request"
-    PIPELINE = "pipeline"
-    MEMBER = "member"
-    TEAM = "team"
-    PROJECT = "project"
